+++ src/Pkg.jl
@@ -131,7 +131,7 @@ Pkg.add(url="/remote/mycompany/juliapackages/OurPackage") # From path to local g
 Pkg.add(url="https://github.com/Company/MonoRepo", subdir="juliapkgs/Package.jl)") # With subdir
 ```
 
-After the installation of new packages the project will be precompiled. See more at [Project Precompilation](@ref).
+After the installation of new packages the project will be precompiled. See more at [Environment Precompilation](@ref).
 
 See also [`PackageSpec`](@ref), [`Pkg.develop`](@ref).
 """
@@ -153,10 +153,6 @@ Precompile all or specific dependencies of the project in parallel.
     Any packages that have previously errored during precompilation won't be retried in auto mode
     until they have changed. To disable automatic precompilation set `ENV["JULIA_PKG_PRECOMPILE_AUTO"]=0`.
     To manually control the number of tasks used set `ENV["JULIA_NUM_PRECOMPILE_TASKS"]`.
-
-!!! compat "Julia 1.3"
-    This function requires at least Julia 1.3. On earlier versions
-    you can use `Pkg.API.precompile()` or the `precompile` Pkg REPL command.
 
 !!! compat "Julia 1.8"
     Specifying packages to precompile requires at least Julia 1.8.
@@ -200,7 +196,7 @@ const why = API.why
     Pkg.update(pkg::Union{String, Vector{String}})
     Pkg.update(pkg::Union{PackageSpec, Vector{PackageSpec}})
 
-If no posistional argument is given, update all packages in the manifest if `mode` is `PKGMODE_MANIFEST` and packages in both manifest and project if `mode` is `PKGMODE_PROJECT`.
+If no positional argument is given, update all packages in the manifest if `mode` is `PKGMODE_MANIFEST` and packages in both manifest and project if `mode` is `PKGMODE_PROJECT`.
 If no positional argument is given, `level` can be used to control by how much packages are allowed to be upgraded (major, minor, patch, fixed).
 
 If packages are given as positional arguments, the `preserve` argument can be used to control what other packages are allowed to update:
@@ -208,7 +204,7 @@ If packages are given as positional arguments, the `preserve` argument can be us
 - `PRESERVE_DIRECT`: Only allow `pkg` and indirect dependencies that are not a direct dependency in the project to update.
 - `PRESERVE_NONE`: Allow `pkg` and all its indirect dependencies to update.
 
-After any package updates the project will be precompiled. See more at [Project Precompilation](@ref).
+After any package updates the project will be precompiled. See more at [Environment Precompilation](@ref).
 
 See also [`PackageSpec`](@ref), [`PackageMode`](@ref), [`UpgradeLevel`](@ref).
 """
@@ -224,9 +220,6 @@ const update = API.up
   - `allow_reresolve::Bool=true`: allow Pkg to reresolve the package versions in the test environment
   - `julia_args::Union{Cmd, Vector{String}}`: options to be passed the test process.
   - `test_args::Union{Cmd, Vector{String}}`: test arguments (`ARGS`) available in the test process.
-
-!!! compat "Julia 1.3"
-    `julia_args` and `test_args` requires at least Julia 1.3.
 
 !!! compat "Julia 1.9"
     `allow_reresolve` requires at least Julia 1.9.
@@ -356,31 +349,34 @@ const generate = API.generate
 """
     Pkg.dependencies()::Dict{UUID, PackageInfo}
 
-!!! compat "Julia 1.4"
-    This feature requires Julia 1.4, and is considered experimental.
+This feature is considered experimental.
 
-Query the dependency graph.
+Query the dependency graph of the active project.
 The result is a `Dict` that maps a package UUID to a `PackageInfo` struct representing the dependency (a package).
 
 # `PackageInfo` fields
 
-| Field             | Description                                                |
-|:------------------|:-----------------------------------------------------------|
-| `name`            | The name of the package                                    |
-| `version`         | The version of the package (this is `Nothing` for stdlibs) |
-| `is_direct_dep`   | The package is a direct dependency                         |
-| `is_tracking_path`| Whether a package is directly tracking a directory         |
-| `is_pinned`       | Whether a package is pinned                                |
-| `source`          | The directory containing the source code for that package  |
-| `dependencies`    | The dependencies of that package as a vector of UUIDs      |
+| Field                 | Description                                                                       |
+|:------------------    |:----------------------------------------------------------------------------------|
+| `name`                | The name of the package                                                           |
+| `version`             | The version of the package (this is `Nothing` for stdlibs)                        |
+| `tree_hash`           | A file hash of the package directory tree                                         |
+| `is_direct_dep`       | The package is a direct dependency                                                |
+| `is_pinned`           | Whether a package is pinned                                                       |
+| `is_tracking_path`    | Whether a package is tracking a path                                              |
+| `is_tracking_repo`    | Whether a package is tracking a repository                                        |
+| `is_tracking_registry`| Whether a package is being tracked by registry i.e. not by path nor by repository |
+| `git_revision`        | The git revision when tracking by repository                                      |
+| `git_source`          | The git source when tracking by repository                                        |
+| `source`              | The directory containing the source code for that package                         |
+| `dependencies`        | The dependencies of that package as a vector of UUIDs                             |
 """
 const dependencies = API.dependencies
 
 """
     Pkg.project()::ProjectInfo
 
-!!! compat "Julia 1.4"
-    This feature requires Julia 1.4, and is considered experimental.
+This feature is considered experimental.
 
 Request a `ProjectInfo` struct which contains information about the active project.
 
@@ -391,6 +387,7 @@ Request a `ProjectInfo` struct which contains information about the active proje
 | name         | The project's name                                                                          |
 | uuid         | The project's UUID                                                                          |
 | version      | The project's version                                                                       |
+| ispackage    | Whether the project is a package (has a name and uuid)                                      |
 | dependencies | The project's direct dependencies as a `Dict` which maps dependency name to dependency UUID |
 | path         | The location of the project file which defines the active project                           |
 """
@@ -409,7 +406,7 @@ If no `Project.toml` exist in the current active project, create one with all th
 dependencies in the manifest and instantiate the resulting project.
 
 After packages have been installed the project will be precompiled.
-See more at [Project Precompilation](@ref).
+See more at [Environment Precompilation](@ref).
 """
 const instantiate = API.instantiate
 
@@ -458,13 +455,6 @@ the output to the difference as compared to the last git commit.
 
 See [`Pkg.project`](@ref) and [`Pkg.dependencies`](@ref) to get the project/manifest
 status as a Julia object instead of printing it.
-
-!!! compat "Julia 1.1"
-    `Pkg.status` with package arguments requires at least Julia 1.1.
-
-!!! compat "Julia 1.3"
-    The `diff` keyword argument requires at least Julia 1.3. In earlier versions `diff=true`
-    is the default for environments in git repositories.
 
 !!! compat "Julia 1.8"
     The `⌃` and `⌅` indicators were added in Julia 1.8.
@@ -515,9 +505,6 @@ Pkg.activate("local/path")
 Pkg.activate("MyDependency")
 Pkg.activate(; temp=true)
 ```
-
-!!! compat "Julia 1.4"
-    the `temp` option requires at least Julia 1.4.
 """
 const activate = API.activate
 
@@ -532,9 +519,6 @@ versions that are already downloaded in version resolution.
 
 To work in offline mode across Julia sessions you can
 set the environment variable `JULIA_PKG_OFFLINE` to `"true"`.
-
-!!! compat "Julia 1.5"
-    Pkg's offline mode requires Julia 1.5 or later.
 """
 offline(b::Bool=true) = (OFFLINE_MODE[] = b; nothing)
 
@@ -569,15 +553,14 @@ This includes:
 Most functions in Pkg take a `Vector` of `PackageSpec` and do the operation on all the packages
 in the vector.
 
-!!! compat "Julia 1.5"
-    Many functions that take a `PackageSpec` or a `Vector{PackageSpec}` can be called with a more concise notation with `NamedTuple`s.
-    For example, `Pkg.add` can be called either as the explicit or concise versions as:
+Many functions that take a `PackageSpec` or a `Vector{PackageSpec}` can be called with a more concise notation with `NamedTuple`s.
+For example, `Pkg.add` can be called either as the explicit or concise versions as:
 
-    | Explicit                                                            | Concise                                        |
-    |:--------------------------------------------------------------------|:-----------------------------------------------|
-    | `Pkg.add(PackageSpec(name="Package"))`                              | `Pkg.add(name = "Package")`                    |
-    | `Pkg.add(PackageSpec(url="www.myhost.com/MyPkg")))`                 | `Pkg.add(name = "Package")`                    |
-    |` Pkg.add([PackageSpec(name="Package"), PackageSpec(path="/MyPkg"])` | `Pkg.add([(;name="Package"), (;path="MyPkg")])`|
+| Explicit                                                            | Concise                                        |
+|:--------------------------------------------------------------------|:-----------------------------------------------|
+| `Pkg.add(PackageSpec(name="Package"))`                              | `Pkg.add(name = "Package")`                    |
+| `Pkg.add(PackageSpec(url="www.myhost.com/MyPkg")))`                 | `Pkg.add(name = "Package")`                    |
+|` Pkg.add([PackageSpec(name="Package"), PackageSpec(path="/MyPkg"])` | `Pkg.add([(;name="Package"), (;path="MyPkg")])`|
 
 Below is a comparison between the REPL mode and the functional API:
 
@@ -589,7 +572,7 @@ Below is a comparison between the REPL mode and the functional API:
 | `Package#master`     | `PackageSpec(name="Package", rev="master")`           |
 | `local/path#feature` | `PackageSpec(path="local/path"; rev="feature")`       |
 | `www.mypkg.com`      | `PackageSpec(url="www.mypkg.com")`                    |
-| `--major Package`    | `PackageSpec(name="Package", version=PKGLEVEL_MAJOR)` |
+| `--major Package`    | `PackageSpec(name="Package", version=UPLEVEL_MAJOR)` |
 
 """
 const PackageSpec = Types.PackageSpec
@@ -639,9 +622,6 @@ A `RegistrySpec` is a representation of a registry with various metadata, much l
 
 Most registry functions in Pkg take a `Vector` of `RegistrySpec` and do the operation
 on all the registries in the vector.
-
-!!! compat "Julia 1.1"
-    Pkg's registry handling requires at least Julia 1.1.
 
 # Examples
 
@@ -702,7 +682,7 @@ function installed()
     for (uuid, dep) in deps
         dep.is_direct_dep || continue
         dep.version === nothing && continue
-        installs[dep.name] = dep.version
+        installs[dep.name] = dep.version::VersionNumber
     end
     return installs
 end